Styling and fixing the log in section
JS: Add an id (id="habitBox") Helped me to style the habit list component.

CSS:Stylings. Adding full width on input fields and on buttons on the log in section.
Add padding on the Log in section.
Styling on the main page and on the habit list component.

HTML: Change the copy on the Log in page and the main page.

Change on the copy on the buttons